// src/models/Stats.ts
export interface weeklyGameStats {
  opponent: string
  week: string
  home: string
  away: string
  completions: string
  attempts: string
  yards: string
  tds: string
  ints: string
  sacks: string
  qbr: string
}

export interface seasonStats {
  completions: string
  attempts: string
  yards: string
  tds: string
  ints: string
  sacks: string
  qbr: string
}